requirements.txt
# Core ML/DL
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
opencv-python>=4.8.0
pillow>=10.0.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Utilities
tqdm>=4.65.0
loguru>=0.7.0

# Testing
pytest>=7.4.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
